edit 'src/scripts/build**' - moved 'buildSync' into own script file
